refactor(data-table): co-locate DataTableSkeleton with DataTable
