+++ docs/src/COMPONENT_API.json
@@ -2710,6 +2710,94 @@
       "rest_props": { "type": "InlineComponent", "name": "TableContainer" }
     },
     {
+      "moduleName": "DataTableSkeleton",
+      "filePath": "/src/DataTable/DataTableSkeleton.svelte",
+      "props": [
+        {
+          "name": "columns",
+          "kind": "let",
+          "description": "Specify the number of columns\nSuperseded by `headers` if `headers` is a non-empty array",
+          "type": "number",
+          "value": "5",
+          "isFunction": false,
+          "constant": false,
+          "reactive": false
+        },
+        {
+          "name": "rows",
+          "kind": "let",
+          "description": "Specify the number of rows",
+          "type": "number",
+          "value": "5",
+          "isFunction": false,
+          "constant": false,
+          "reactive": false
+        },
+        {
+          "name": "size",
+          "kind": "let",
+          "description": "Set the size of the data table",
+          "type": "\"compact\" | \"short\" | \"tall\"",
+          "isFunction": false,
+          "constant": false,
+          "reactive": false
+        },
+        {
+          "name": "zebra",
+          "kind": "let",
+          "description": "Set to `true` to apply zebra styles to the datatable rows",
+          "type": "boolean",
+          "value": "false",
+          "isFunction": false,
+          "constant": false,
+          "reactive": false
+        },
+        {
+          "name": "showHeader",
+          "kind": "let",
+          "description": "Set to `false` to hide the header",
+          "type": "boolean",
+          "value": "true",
+          "isFunction": false,
+          "constant": false,
+          "reactive": false
+        },
+        {
+          "name": "headers",
+          "kind": "let",
+          "description": "Set the column headers\nSupersedes `columns` if value is a non-empty array",
+          "type": "string[] | Partial<DataTableHeader>[]",
+          "value": "[]",
+          "isFunction": false,
+          "constant": false,
+          "reactive": false
+        },
+        {
+          "name": "showToolbar",
+          "kind": "let",
+          "description": "Set to `false` to hide the toolbar",
+          "type": "boolean",
+          "value": "true",
+          "isFunction": false,
+          "constant": false,
+          "reactive": false
+        }
+      ],
+      "slots": [],
+      "events": [
+        { "type": "forwarded", "name": "click", "element": "table" },
+        { "type": "forwarded", "name": "mouseover", "element": "table" },
+        { "type": "forwarded", "name": "mouseenter", "element": "table" },
+        { "type": "forwarded", "name": "mouseleave", "element": "table" }
+      ],
+      "typedefs": [],
+      "rest_props": { "type": "Element", "name": "div" },
+      "extends": {
+        "interface": "DataTableHeader",
+        "import": "\"../DataTable/DataTable\""
+      }
+    },
+    {
       "moduleName": "Toolbar",
       "filePath": "/src/DataTable/Toolbar.svelte",
       "props": [
@@ -3314,94 +3402,6 @@
       "extends": {
         "interface": "OverflowMenuItemProps",
         "import": "\"../OverflowMenu/OverflowMenuItem\""
-      }
-    },
-    {
-      "moduleName": "DataTableSkeleton",
-      "filePath": "/src/DataTableSkeleton/DataTableSkeleton.svelte",
-      "props": [
-        {
-          "name": "columns",
-          "kind": "let",
-          "description": "Specify the number of columns\nSuperseded by `headers` if `headers` is a non-empty array",
-          "type": "number",
-          "value": "5",
-          "isFunction": false,
-          "constant": false,
-          "reactive": false
-        },
-        {
-          "name": "rows",
-          "kind": "let",
-          "description": "Specify the number of rows",
-          "type": "number",
-          "value": "5",
-          "isFunction": false,
-          "constant": false,
-          "reactive": false
-        },
-        {
-          "name": "size",
-          "kind": "let",
-          "description": "Set the size of the data table",
-          "type": "\"compact\" | \"short\" | \"tall\"",
-          "isFunction": false,
-          "constant": false,
-          "reactive": false
-        },
-        {
-          "name": "zebra",
-          "kind": "let",
-          "description": "Set to `true` to apply zebra styles to the datatable rows",
-          "type": "boolean",
-          "value": "false",
-          "isFunction": false,
-          "constant": false,
-          "reactive": false
-        },
-        {
-          "name": "showHeader",
-          "kind": "let",
-          "description": "Set to `false` to hide the header",
-          "type": "boolean",
-          "value": "true",
-          "isFunction": false,
-          "constant": false,
-          "reactive": false
-        },
-        {
-          "name": "headers",
-          "kind": "let",
-          "description": "Set the column headers\nSupersedes `columns` if value is a non-empty array",
-          "type": "string[] | Partial<DataTableHeader>[]",
-          "value": "[]",
-          "isFunction": false,
-          "constant": false,
-          "reactive": false
-        },
-        {
-          "name": "showToolbar",
-          "kind": "let",
-          "description": "Set to `false` to hide the toolbar",
-          "type": "boolean",
-          "value": "true",
-          "isFunction": false,
-          "constant": false,
-          "reactive": false
-        }
-      ],
-      "slots": [],
-      "events": [
-        { "type": "forwarded", "name": "click", "element": "table" },
-        { "type": "forwarded", "name": "mouseover", "element": "table" },
-        { "type": "forwarded", "name": "mouseenter", "element": "table" },
-        { "type": "forwarded", "name": "mouseleave", "element": "table" }
-      ],
-      "typedefs": [],
-      "rest_props": { "type": "Element", "name": "div" },
-      "extends": {
-        "interface": "DataTableHeader",
-        "import": "\"../DataTable/DataTable\""
       }
     },
     {
