--- src/DataTableSkeleton/index.js
@@ -1,1 +1,0 @@
-export { default as DataTableSkeleton } from "./DataTableSkeleton.svelte";
