+++ types/index.d.ts
@@ -39,6 +39,7 @@ export { default as TableHead } from "./DataTable/TableHead";
 export { default as TableHeader } from "./DataTable/TableHeader";
 export { default as TableRow } from "./DataTable/TableRow";
 export { default as DataTable } from "./DataTable/DataTable";
+export { default as DataTableSkeleton } from "./DataTable/DataTableSkeleton";
 export { default as Toolbar } from "./DataTable/Toolbar";
 export { default as ToolbarContent } from "./DataTable/ToolbarContent";
 export { default as SearchSkeleton } from "./Search/SearchSkeleton";
@@ -49,7 +50,6 @@ export { default as OverflowMenu } from "./OverflowMenu/OverflowMenu";
 export { default as OverflowMenuItem } from "./OverflowMenu/OverflowMenuItem";
 export { default as ToolbarMenu } from "./DataTable/ToolbarMenu";
 export { default as ToolbarMenuItem } from "./DataTable/ToolbarMenuItem";
-export { default as DataTableSkeleton } from "./DataTableSkeleton/DataTableSkeleton";
 export { default as DatePicker } from "./DatePicker/DatePicker";
 export { default as DatePickerInput } from "./DatePicker/DatePickerInput";
 export { default as DatePickerSkeleton } from "./DatePicker/DatePickerSkeleton";
